requirements.txt
asgiref==3.9.2
certifi==2025.10.5
charset-normalizer==3.4.3
Django==5.2.6
django-widget-tweaks==1.5.0
idna==3.10
psycopg2-binary==2.9.10
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
